Major redesign
Some highlights:
- Return structures uniformized, removing duplicated functions just with
different structures as arguments.
- Major redesign on 'Finder'.
- Changed names of variables and functions.
- 'Finder' functions became non-static.
- On option 8 (list dir1 files with copies on dir2), progress was
showing the number of copies found, but the result was showing the
number of originals with copies - this bug is fixed with this update.
- Arguments of 'findFiles1WithCopiesSomewhereInFiles2' rotated and
renamed relatively to its old version 'findCopiesOfFilesToSearch'.
- Minor improvements on 'findDuplicates'.
- Text related to the variable
'dir1FilesWithoutCorrespondentInDir2ButWithCopiesThereMsg' of 'Text'
updated to contain the number of files found.
- Text related to the variable 'optionsMsg' of 'Text' updated to be
easier to understand.