HashMap changed to LinkedHashMap
HashMap changed to LinkedHashMap to preserve the insertion order (file
tree loading order) instead of using the hash code order while the Map
is iterated. This way, the list sent to the log is better ordered. Due
to this change, failures in the calculation of the remaining time are
more perceptible.