Remaining time calculation improved
RemainingTime is now receiving the number of bytes already processed and
the total of bytes to process, so the measure is uniform. It was
receiving the number of files already processed and the total of files
to process which was not a uniform measure because each file has its own
length.